Fix ability distinction and add autoclick support
Apparently Unity cannot distinguish between multiple components of the
same type on the same GameObject when attempting to assign them to
inspector object fields (only the first such component is always
assigned to the field). The prefab structure for the ability system was
predicated on it being able to do this. Now each ability component must
exist on a child object of the combat entity instead of on the same
GameObject.